Refactor wiki homepage template lookup.
Because the parent BP Docs plugin implemented theme compat mode, we also
want to make the change to use a template part.

However, we also want to be considerate to older users of the plugin that
had access to the specialized wiki sidebar widget.  Implementing theme
compat would remove the wiki sidebar widget.

Thus, to maintain backwards compatibility for older users, the plugin now
checks for a 'docs/index-wiki.php' template in the theme.

If the current theme has this template, the plugin will use this.

If the theme does not have this template, theme compat will be used.  A
new template part is used for theme compat - 'docs/home-wiki.php'.

Fixes #8.