When setting up wiki-only tags, fetch all wiki tags in one fell swoop
When querying for wiki items (in order to filter only by wiki tags),
posts_per_page is set to -1. This prevents WP from pre-fetching term data.
This means that the individual calls to wp_get_object_terms() were going
directly to the database, creating large overhead on sites with a large number
of wiki items.

This changeset rearranges the wiki term query so that they are pre-fetched in
a single query (using update_post_caches()) and then grabbed directly out of
the object cache (because wp_get_object_terms() does not look at the term cache
for some odd reason).